Move partitioning.
